Move interfaces to separate catalog in grand.api project
