+++ Grand.Api/Grand.Api.csproj
@@ -16,6 +16,6 @@
   </ItemGroup>
   <ItemGroup>
     <PackageReference Include="Microsoft.AspNetCore.OData" Version="7.1.0" />
-  </ItemGroup>
+  </ItemGroup>  
 
 </Project>
