+++ Grand.Api/Interfaces/ICommonApiService.cs
@@ -1,7 +1,7 @@
 ﻿using Grand.Api.DTOs.Common;
 using MongoDB.Driver.Linq;
 
-namespace Grand.Api.Services
+namespace Grand.Api.Interfaces
 {
     public interface ICommonApiService
     {
