+++ Grand.Api/Interfaces/ICustomerRoleApiService.cs
@@ -1,7 +1,7 @@
 ﻿using Grand.Api.DTOs.Customers;
 using MongoDB.Driver.Linq;
 
-namespace Grand.Api.Services
+namespace Grand.Api.Interfaces
 {
     public interface ICustomerRoleApiService
     {
