+++ Grand.Api/Interfaces/IShippingApiService.cs
@@ -1,7 +1,7 @@
 ﻿using Grand.Api.DTOs.Shipping;
 using MongoDB.Driver.Linq;
 
-namespace Grand.Api.Services
+namespace Grand.Api.Interfaces
 {
     public interface IShippingApiService
     {
