+++ Grand.Api/Services/CustomerApiService.cs
@@ -1,5 +1,6 @@
 ﻿using Grand.Api.DTOs.Customers;
 using Grand.Api.Extensions;
+using Grand.Api.Interfaces;
 using Grand.Core.Domain.Customers;
 using Grand.Data;
 using Grand.Services.Common;
