+++ Grand.Api/Services/CustomerRoleApiService.cs
@@ -1,5 +1,6 @@
 ﻿using Grand.Api.DTOs.Customers;
 using Grand.Api.Extensions;
+using Grand.Api.Interfaces;
 using Grand.Data;
 using Grand.Services.Customers;
 using Grand.Services.Localization;
