+++ Grand.Web/Areas/Api/Controllers/OData/ManufacturerTemplateController.cs
@@ -1,4 +1,4 @@
-﻿using Grand.Api.Services;
+﻿using Grand.Api.Interfaces;
 using Grand.Services.Security;
 using Microsoft.AspNet.OData;
 using Microsoft.AspNetCore.Mvc;
