+++ Tests/Grand.Api.Tests/ControllerTests/CategoryControllerTest.cs
@@ -1,4 +1,5 @@
 ﻿using Grand.Api.DTOs.Catalog;
+using Grand.Api.Interfaces;
 using Grand.Api.Services;
 using Grand.Api.Tests.Helpers;
 using Grand.Core.Data;
