+++ Tests/Grand.Api.Tests/ServicesTests/CategoryApiServiceTest.cs
@@ -1,4 +1,5 @@
-﻿using Grand.Api.Services;
+﻿using Grand.Api.Interfaces;
+using Grand.Api.Services;
 using Grand.Api.Tests.Helpers;
 using Grand.Core.Data;
 using Grand.Core.Domain.Catalog;
